+++ src/contexts/index.ts
@@ -1,1 +1,1 @@
-export * from './use-device-size'
+export * from './device-size'
